refactor(common-tests): extract ComponentDepthFinder test helper
Both `OrganizersSystemsTests` and `OrganizersSystemsEndToEndTests` had
their own copy of the recursive `MeasureDepth` walk used to assert
tree-depth invariants for `Device.AddComponent()` auto-wrap behaviour.
Extract the walk into `TestHelpers/ComponentDepthFinder` so the logic
stays in one place; future depth-based tree assertions consume the
shared helper instead of forking another copy.

The helper exposes both an entry-point overload that takes a `Device`
and an internal overload that takes an `IEnumerable<IComponent>` plus
a starting depth, so callers can either ask "what depth is X under
this device?" or weave the walk into a custom traversal.
